+++ backend/canisters/claimlink/api/src/queries/mod.rs
@@ -9,6 +9,8 @@ pub mod get_mint_request;
 pub mod get_mint_requests_by_owner;
 pub mod get_nft_details;
 pub mod get_ogy_usd_price;
+pub mod get_template_by_id;
+pub mod get_template_ids_by_owner;
 pub mod get_templates_by_owner;
 pub mod http_request;
 pub mod icrc_10;